requirements.txt
pandas>=2.0
pyarrow>=14.0
numpy>=1.24
matplotlib>=3.7
scikit-learn>=1.3
requests>=2.28
joblib>=1.3
jupyter>=1.0
tqdm>=4.65
python-dotenv>=1.0
pytest>=7.0
shap>=0.43
